Per #1484, split vx_util/get_filenames.cc into 2 pieces and move file list parsing functionality into vx_data2d_factory/parse_file_list.cc. Had to update includes for tc_gen and gsidens2orank to handle this change. Next, I need to tweak the parse_file_list logic slightly.
